+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Added `terminal?` function
+
 # 0.1.0
 
 * Initial release